issue #9 - Make a simplified OpenCloseTrait, and a more complicated MultiDirectionalOpenCloseTrait as per
