+++ command.go
@@ -26,11 +26,11 @@ func (t OnOffCommand) Name() string {
 	return "action.devices.commands.OnOff"
 }
 
-// OpenCloseCommand defines how a function should handle this specific trait
-type OpenCloseCommand func(ctx Context, openPercent float64, openDirection OpenCloseTraitDirection) proto.DeviceError
+// DirectionalOpenCloseCommand defines how a function should handle this specific trait
+type DirectionalOpenCloseCommand func(ctx Context, openPercent float64, openDirection OpenCloseTraitDirection) proto.DeviceError
 
 // Execute validates the request and calls the user defined handler for the device trait
-func (t OpenCloseCommand) Execute(ctx Context, args map[string]interface{}) proto.DeviceError {
+func (t DirectionalOpenCloseCommand) Execute(ctx Context, args map[string]interface{}) proto.DeviceError {
 	// validate if our EXECUTE Request contains the "openPercent" object so the handler can actually handle something
 	openDirection := OpenCloseTraitDirectionNone
 	if argOpenPercent, ok := args["openPercent"]; ok {
@@ -42,6 +42,26 @@ func (t OpenCloseCommand) Execute(ctx Context, args map[string]interface{}) prot
 				}
 			}
 			return t(ctx, openPercent, openDirection)
+		}
+		return proto.ErrorCodeNotSupported
+	}
+	return proto.ErrorCodeProtocolError
+}
+
+// Name returns the intent for this Trait
+func (t DirectionalOpenCloseCommand) Name() string {
+	return "action.devices.commands.OpenClose"
+}
+
+// OpenCloseCommand defines how a function should handle this specific trait
+type OpenCloseCommand func(ctx Context, openPercent float64) proto.DeviceError
+
+// Execute validates the request and calls the user defined handler for the device trait
+func (t OpenCloseCommand) Execute(ctx Context, args map[string]interface{}) proto.DeviceError {
+	// validate if our EXECUTE Request contains the "openPercent" object so the handler can actually handle something
+	if argOpenPercent, ok := args["openPercent"]; ok {
+		if openPercent, ok := argOpenPercent.(float64); ok {
+			return t(ctx, openPercent)
 		}
 		return proto.ErrorCodeNotSupported
 	}
